wire up global cost filters
