+++ gui/src/lib/app/filter-dialog/global-filter-form/global-filter-form.svelte
@@ -1,22 +1,5 @@
 <script lang="ts">
-    import type { CostTier } from '$lib/app/form-context/types'
-    import GoldIcon from '$lib/icons/gold-icon.svelte'
-    import { range } from 'radash'
-    import { getFilterFormContext } from '../../form-context/context'
-    import CheckboxGroup, {
-        type CheckboxData
-    } from '../checkbox-group/checkbox-group.svelte'
-
-    const { form, controls } = getFilterFormContext()
-
-    const options = (Array.from(range(1, 5)) as CostTier[]).map(
-        (idx) => ({
-            value: true,
-            label: idx.toString(),
-            suffix: GoldIcon,
-            onChange: (update: boolean) => {}
-        })
-    ) satisfies CheckboxData[]
+    import GlobalCostInput from './global-cost-input.svelte'
 </script>
 
 <form class="h-full w-full pt-4 px-6">
@@ -24,9 +7,7 @@
         Global Filters
     </h1>
 
-    <div>
-        <CheckboxGroup label="Cost" {options} />
-    </div>
+    <GlobalCostInput />
 </form>
 
 <style lang="postcss">
